폴더구조를 정했습니다git add .짝짝
