--- src/components/pages/Layout.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Layout = ({ children }) => {
-  return (
-    <div>
-      <h5>Layout</h5>
-      {children}
-    </div>
-  );
-};
-
-export default Layout;
