--- src/components/pages/Post.js
@@ -1,16 +1,0 @@
-import Main from "../templates/Main";
-import PostForm from "../../UI/molecules/PostForm";
-import Header from "../templates/Header";
-
-function Post() {
-  return (
-    <>
-      <Header />
-      <Main>
-        <PostForm />
-      </Main>
-    </>
-  );
-}
-
-export default Post;
