+++ src/components/templates/Header.js
@@ -1,8 +1,13 @@
+import { useNavigate } from "react-router-dom";
+
 const Header = () => {
+  const navi = useNavigate();
   return (
-    <>
-      <div>헤더입니다.</div>
-    </>
+    <header>
+      <button onClick={() => navi("/")}> 홈으로</button>
+      <h1>개발로그9조쓰</h1>
+      <button onClick={() => navi(-1)}>뒤로가기</button>
+    </header>
   );
 };
 export default Header;
