+++ src/components/templates/Layout.js
@@ -1,0 +1,17 @@
+import React from "react";
+import styled from "styled-components";
+
+const Layout = ({ children }) => {
+  return <StLayout>{children}</StLayout>;
+};
+
+export default Layout;
+
+const StLayout = styled.div`
+  max-width: 1200px;
+  min-width: 800px;
+
+  min-height: 100vh;
+
+  margin: 0 auto;
+`;
