--- src/components/templates/Main.js
@@ -1,8 +1,0 @@
-const Main = (props) => {
-  return (
-    <>
-      <div>{props.children}</div>
-    </>
-  );
-};
-export default Main;
