+++ src/shared/Router.js
@@ -1,21 +1,17 @@
 import React from "react";
-import { Route, Routes, Link } from "react-router-dom";
-import Home from "../components/pages/Home";
-import Commit from "../components/pages/Commit";
-import Record from "../components/pages/Record";
-import Info from "../components/pages/Info";
+import { Route, Routes } from "react-router-dom";
+
+import HomePage from "../components/pages/HomePage";
+import PostPage from "../components/pages/PostPage";
+import DetailPage from "../components/pages/DetailPage";
 
 const Router = () => {
   return (
-    <div>
-      <h1>Router</h1>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/commit" element={<Commit />} />
-        <Route path="/record" element={<Record />} />
-        <Route path="/info" element={<Info />} />
-      </Routes>
-    </div>
+    <Routes>
+      <Route path="/" element={<HomePage />} />
+      <Route path="/post" element={<PostPage />} />
+      <Route path="/detail" element={<DetailPage />} />
+    </Routes>
   );
 };
